+++ autogpt_platform/backend/test/test_data_creator.py
@@ -10,16 +10,12 @@ from prisma.types import (
     AgentGraphCreateInput,
     AgentNodeCreateInput,
     AgentNodeLinkCreateInput,
-    AgentPresetCreateInput,
     AnalyticsDetailsCreateInput,
     AnalyticsMetricsCreateInput,
     APIKeyCreateInput,
     CreditTransactionCreateInput,
-    LibraryAgentCreateInput,
     ProfileCreateInput,
-    StoreListingCreateInput,
     StoreListingReviewCreateInput,
-    StoreListingVersionCreateInput,
     UserCreateInput,
 )
